requirements.txt
numpy
scikit-learn
chromadb
streamlit
torch
torchvision
Pillow
transformers
